Move definition to its own file
