+++ lib/graphql/client.rb
@@ -3,6 +3,7 @@ require "active_support/inflector"
 require "active_support/notifications"
 require "graphql"
 require "graphql/client/collocated_enforcement"
+require "graphql/client/definition"
 require "graphql/client/definition_variables"
 require "graphql/client/error"
 require "graphql/client/errors"
@@ -86,96 +87,6 @@ module GraphQL
       @document_tracking_enabled = false
       @allow_dynamic_queries = false
       @enforce_collocated_callers = enforce_collocated_callers
-    end
-
-    # Definitions are constructed by Client.parse and wrap a parsed AST of the
-    # query string as well as hold references to any external query definition
-    # dependencies.
-    #
-    # Definitions MUST be assigned to a constant.
-    class Definition < Module
-      def self.for(node:, **kargs)
-        case node
-        when Language::Nodes::OperationDefinition
-          OperationDefinition.new(node: node, **kargs)
-        when Language::Nodes::FragmentDefinition
-          FragmentDefinition.new(node: node, **kargs)
-        else
-          raise TypeError, "expected node to be a definition type, but was #{node.class}"
-        end
-      end
-
-      def initialize(node:, document:, schema:, document_types:, source_location:, enforce_collocated_callers:)
-        @definition_node = node
-        @document = document
-        @schema = schema
-        @document_types = document_types
-        @source_location = source_location
-        @enforce_collocated_callers = enforce_collocated_callers
-      end
-
-      # Internal: Get underlying operation or fragment defintion AST node for
-      # definition.
-      #
-      # Returns OperationDefinition or FragmentDefinition object.
-      attr_reader :definition_node
-
-      # Public: Global name of definition in client document.
-      #
-      # Returns a GraphQL safe name of the Ruby constant String.
-      #
-      #   "Users::UserQuery" #=> "Users__UserQuery"
-      #
-      # Returns String.
-      def definition_name
-        return @definition_name if defined?(@definition_name)
-
-        if name
-          @definition_name = name.gsub("::", "__").freeze
-        else
-          "#{self.class.name}_#{object_id}".gsub("::", "__").freeze
-        end
-      end
-
-      # Public: Get document with only the definitions needed to perform this
-      # operation.
-      #
-      # Returns GraphQL::Language::Nodes::Document with one OperationDefinition
-      # and any FragmentDefinition dependencies.
-      attr_reader :document
-
-      # Internal: Mapping of document nodes to schema types.
-      attr_reader :document_types
-
-      attr_reader :schema
-
-      # Public: Returns the Ruby source filename and line number containing this
-      # definition was not defined in Ruby.
-      #
-      # Returns Array pair of [String, Fixnum].
-      attr_reader :source_location
-
-      attr_reader :enforce_collocated_callers
-
-      def new(*args)
-        type.new(*args)
-      end
-
-      def type
-        # TODO: Fix type indirection
-        @type ||= GraphQL::Client::QueryResult.wrap(self, definition_node, name: "#{name}.type")
-      end
-    end
-
-    # Specific operation definition subtype for queries, mutations or
-    # subscriptions.
-    class OperationDefinition < Definition
-      # Public: Alias for definition name.
-      alias operation_name definition_name
-    end
-
-    # Specific fragment definition subtype.
-    class FragmentDefinition < Definition
     end
 
     def parse(str, filename = nil, lineno = nil)
